requirements.txt
streamlit == 1.39.0
matplotlib==3.9.2
pandas == 2.2.3
python-dotenv == 1.0.0
streamlit-option-menu==0.4.0
plotly==5.24.1